spring-mockito/src/main/java/com/example/mockito/user/UserService.java
package com.example.mockito.user;

import com.example.mockito.user.repository.UserRepository;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.stereotype.Service;

@Service
public class UserService {

    private final UserRepository userRepository;

    public UserService(@Qualifier("defaultUserRepository") UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public User getUser(Long id) {
        return userRepository.findUser(id);
    }
}
